+++ src/spring-cloud/HISTORY.md
@@ -1,8 +1,17 @@
 Release History
 ===============
-1.2.1
+
+2.0.2
 -----
 * Add optional '--deployment' to 'az spring-cloud app logs' command
+
+2.0.1
+-----
+* Fix 'az spring-cloud app list' command issues.
+
+2.0.0
+-----
+* Switch api-version from 2019-05-01-preview to 2020-07-01
 
 1.2.0
 -----